Move stats to their own cog

+++ dm_assist/db/alembic/server/versions/9d6b3231a6fe_add_rollstats.py
@@ -25,7 +25,7 @@ def upgrade():
     )
 
     with op.batch_alter_table('stats') as batch_op:
-        batch_op.add_column(sa.Column('calc', sa.Integer()))
+        batch_op.add_column(sa.Column('calc', sa.Float()))
 
 
 def downgrade():
